Remove unnecessary user authentication and contact message files
This commit removes multiples files related to user authentication and contact messages. These files include entity files for user requests, controllers and services for user authentication, utilities for jwt token validation and exceptions for bad request and conflict. It also eliminates payload requests and responses along with corresponding mappers and validators. This cleanup reduces confusion and redundancy in the codebase.