Add admin update feature and user existence check in UserService
Implemented a feature for admins to update Dean and Vice Dean in UserController. Added a method in MethodHelper to check user existence to minimize exceptions due to non-existent users. Adjusted the getUser method in UserService to get the "username" value from the request header instead of its attribute. This change introduces more direct error handling and expands admin functionality.