Improve user handling in UserController and UserService
The update includes the addition of a feature in UserController that allows admins to update Dean and ViceDean. A method has been added in MethodHelper to verify user existence and reduce the occurrence of exceptions due to non-existent users. The getUser method in UserService has been revised to extract the "username" value directly from the request header instead of its attribute. These alterations improve error handling and provide more control to admins.