Implement security features and add UserDetailService
The update introduces several security features such as the "UserDetailServiceImpl" and "UserDetailsImpl". The JWT handling utilities have also been added. This includes the generation and validation of JWT tokens and extraction of usernames from these tokens. A method named "loadUserByName" was also added in the "MethodHelper" class, which is used to load a user's details through their username. Dependencies necessary for these features were also made active in the pom.xml file.