Added EducationTerm functionality and updated user authentication
Implemented the full set of necessary files and classes (controller, service, repository, request and response models) for the new EducationTerm domain. This segment includes the management of education terms in a school context. In addition, enhanced the user authentication by introducing a function to find user details by username, and expanded AuthenticationService and AuthenticationController to support this feature.